refactor(agents): extract agent-archive lifecycle into agent_archive.py (B2a)
